% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cache}
\alias{cache}
\alias{sec_cache_dir}
\alias{sec_cache_clear}
\title{Manage the insidertrade cache}
\usage{
sec_cache_dir()

sec_cache_clear()
}
\description{
\code{sec_cache_dir()} returns the path where cached data is stored.
\code{sec_cache_clear()} deletes all cached files, including bulk data ZIPs and API responses.
}
\details{
Bulk data ZIPs from \code{\link[=sec_form345]{sec_form345()}} are always cached on disk. API responses from
\code{\link[=sec_tickers]{sec_tickers()}} and \code{\link[=edgar_submissions]{edgar_submissions()}} are only cached when enabled with
\code{options(insidertrade.cache = TRUE)}. Cached API responses are stored for 7 days by default,
but this can be customized with \code{options(insidertrade.cache_max_age = days)}.
}
\examples{
# view cache location
sec_cache_dir()

\dontrun{
# enable API caching
options(insidertrade.cache = TRUE)

# clear all cached data
sec_cache_clear()
}
}
