% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar.R
\name{edgar_insider_filings}
\alias{edgar_insider_filings}
\title{EDGAR insider filings}
\source{
\url{https://www.sec.gov/search-filings/edgar-application-programming-interfaces}
}
\usage{
edgar_insider_filings(cik)
}
\arguments{
\item{cik}{(\code{integer(1)} | \code{character(1)})\cr
The SEC Central Index Key for the entity.}
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table()}} with one row per insider filing.
}
\description{
A convenience wrapper around \code{\link[=edgar_submissions]{edgar_submissions()}} that filters to insider trading forms
(Forms 3, 4, and 5).
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
\donttest{
# Apple Inc. insider filings
insider <- edgar_insider_filings(320193)
}
\dontshow{\}) # examplesIf}
}
\seealso{
Other EDGAR: 
\code{\link{edgar_submissions}()},
\code{\link{sec_tickers}()}
}
\concept{EDGAR}
