% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sec_cache_dir}
\alias{sec_cache_dir}
\title{Get the cache directory}
\usage{
sec_cache_dir()
}
\value{
A \code{character()} string with the path to the cache directory.
}
\description{
Returns the path to the directory where downloaded SEC data files are cached.
}
\seealso{
Other cache: 
\code{\link{sec_cache_clear}()}
}
\concept{cache}
