% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec.R
\name{sec_transaction_codes}
\alias{sec_transaction_codes}
\title{SEC transaction codes}
\source{
\url{https://www.sec.gov/edgar/searchedgar/ownershipformcodes.html}
}
\usage{
sec_transaction_codes()
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table()}} with columns \code{code} and \code{description}.
}
\description{
Returns a lookup table mapping SEC transaction codes to human-readable descriptions.
}
\examples{
sec_transaction_codes()
}
\seealso{
Other bulk data: 
\code{\link{sec_form345}()},
\code{\link{sec_transactions}()}
}
\concept{bulk data}
