% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec.R
\name{sec_transactions}
\alias{sec_transactions}
\title{SEC insider transactions}
\source{
\url{https://www.sec.gov/data-research/sec-markets-data/insider-transactions-data-sets}
}
\usage{
sec_transactions(year, quarter)
}
\arguments{
\item{year}{(\code{integer(1)})\cr
The year of the filing (e.g. \code{2024}).}

\item{quarter}{(\code{integer(1)})\cr
The fiscal quarter. Must be between 1 and 4.}
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table()}} with transaction data joined with submission and reporting owner
information.
}
\description{
Download the quarterly Form 3/4/5 data and join the \code{nonderiv_trans} table with \code{submission}
and \code{reportingowner} tables on \code{accession_number}.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
\donttest{
trans <- sec_transactions(2024, 2)
}
\dontshow{\}) # examplesIf}
}
\seealso{
Other bulk data: 
\code{\link{sec_form345}()}
}
\concept{bulk data}
