% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insidertrade-package.R
\docType{package}
\name{insidertrade-package}
\alias{insidertrade}
\alias{insidertrade-package}
\title{insidertrade: SEC Insider Trading Transactions}
\description{
Download and parse SEC insider trading data (Forms 3/4/5) from quarterly bulk data sets and the EDGAR submissions API.
}
\section{Options}{

The following options configure package behavior:
\itemize{
\item \code{insidertrade.user_agent}: \strong{Required.} A string with your contact information
(e.g. \code{"your@email.com"}). The SEC requires a valid User-Agent for all API requests.
\item \code{insidertrade.cache}: If \code{TRUE}, cache API responses to disk. Default \code{FALSE}.
\item \code{insidertrade.cache_max_age}: Number of days to keep cached API responses. Default \code{7}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://m-muecke.github.io/insidertrade/}
  \item \url{https://github.com/m-muecke/insidertrade}
  \item Report bugs at \url{https://github.com/m-muecke/insidertrade/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Mücke \email{muecke.maximilian@gmail.com} (\href{https://orcid.org/0009-0000-9432-9795}{ORCID})

}
\keyword{internal}
