% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sec_cache_clear}
\alias{sec_cache_clear}
\title{Clear the cache}
\usage{
sec_cache_clear()
}
\value{
\code{NULL} invisibly.
}
\description{
Deletes all cached ZIP files downloaded by \code{\link[=sec_form345]{sec_form345()}}.
}
\seealso{
Other cache: 
\code{\link{sec_cache_dir}()}
}
\concept{cache}
